Make fixes from code review
